mtx: drop redundant (int16_t) casts before 16-bit masks
Answering the user's question about `(uint32_t)(int16_t)` double casts. That
specific combo appears at only 4 sites in the tree, and 3 of them are in
mtx.cpp (introduced by the Trig/Math/Mtx promotion) where a mask to 16 bits
follows immediately — `(int16_t)x & 0xFFFF` and `x & 0xFFFF` produce identical
bits, so the sign-extension is a no-op. Dropped it.

The lone remaining `(uint32_t)(int16_t)` site (beh_camera_target_follow.cpp) is
legitimate — a right-shift follows, and the sign-extension IS what makes the
shift sign-preserving. Not enough repetition to warrant a helper macro/inline.

Related: `(int16_t)X` cast-of-value pattern is ~282 sites tree-wide but almost
all are legitimate MIPS-idiom transcriptions (sign-extract high half, truncate
for mem_w16, float conversion, etc.); no single sub-pattern dominates enough to
factor. mem_r16s/r8s already covered the memory-read case earlier this session.

Clean build.

+++ game/math/mtx.cpp
@@ -17,10 +17,11 @@ void Mtx::identity(uint32_t addr) {
 
 void Mtx::diagonal(uint32_t addr, int32_t x, int32_t y, int32_t z) {
   Core* c = this->core;
-  // Guest sign-extends each arg with sll 16 / sra 16 (drop high half) — reproduce via u16 casts.
-  uint32_t xw = (uint32_t)(int16_t)x & 0xFFFFu;
-  uint32_t yw = (uint32_t)(int16_t)y & 0xFFFFu;
-  uint32_t zw = (uint32_t)(int16_t)z & 0xFFFFu;
+  // Guest drops the high half via sll 16 / sra 16; the mask below is what actually matters (the
+  // sign-extend is a no-op once we `& 0xFFFF`, so no (int16_t) cast needed).
+  uint32_t xw = (uint32_t)x & 0xFFFFu;
+  uint32_t yw = (uint32_t)y & 0xFFFFu;
+  uint32_t zw = (uint32_t)z & 0xFFFFu;
   c->mem_w32(addr +  0, xw);   // m[0][0]=x  m[0][1]=0
   c->mem_w32(addr +  4, 0);    // m[0][2]=0  m[1][0]=0
   c->mem_w32(addr +  8, yw);   // m[1][1]=y  m[1][2]=0
